Organize variables in sub-sections and remove value alignment.
Aligning variables got a bit out of hand. I was aligning every
variable to the same column, so even if it's a short variable
the colon could potentially be really far out.

I've also lately been doing a bunch of work then commiting,
so adding more variables, spacing out the other values and
then looking at my changes for commiting was a confusing mess.

If I'm going to do this again variables will be aligned to others
in their section, and not with every single variable in the file.